v2.5.1.7 -> v2.5.1.8
- Greg KH: USB updates
- various: kdev_t updates
- Al Viro: more bread()/filesystem cleanups

+++ drivers/usb/usbnet.c
@@ -1571,7 +1571,7 @@ static int usbnet_start_xmit (struct sk_buff *skb, struct net_device *net)
 	struct urb		*urb = 0;
 	struct skb_data		*entry;
 	struct driver_info	*info = dev->driver_info;
-	int			flags;
+	unsigned long		flags;
 #ifdef	CONFIG_USB_NET1080
 	struct nc_header	*header = 0;
 	struct nc_trailer	*trailer = 0;
